MOS6551: correctly transfer data from RSR to RDR (#13815)
On real hardware, this transfer is not atomic, but
done bit-by-bit on the last clock cycles of a byte
reception. Do the same, so bugs in programs are not
hidden from developers.

Also fixes another discrepancy with reality: when
the RDR is full and a byte arrives, it is transferred
to the RDR, and the Overrun flag is set. Our
implementation did set the flag but left the old byte
in the RDR.