osd/modules/file: define buffer size for openpty() (#13837)
There is no defined size for the slave name returned by openpty();
OSes that have a PATH_MAX constant usually use that as limit.
Create a custom OPENPTY_PATH_MAX constant specifically for openpty(),
even in case PATH_MAX is not available: in that situation, use 8192 as
buffer size, which hopefully should be "good enough".