posixfile.cpp: remove redundant realpath() call
Commit 1bd116c5c8f8f3d28e4db7925efcb87279ba860b makes use of the
allocating mode of realpath() by duplicating the realpath() invocation
already done (i.e. with the same input path). In practice, if the
realpath() invocation with nullptr fails, the realpath() invocation with
a destination buffer is very likely to fail in the same way.

Hence drop the realpath() invocation with a destination buffer, and move
the realpath() invocation with nullptr in the existing if/else if chain
to maintain the existing behaviour.

Since "path_buffer" now is needed only for the fallback getcwd()
invocation, move it in the inner scope where it is used. Since the
usage of getcwd() already has a growing buffer logic, make the initial
buffer size shorter, which should be enough to get current working
directory already at the first attempt in most of the cases.