Windows:  Do not drop into debugger when no MAME window has focus
On Windows, the debugger has a special Windows-specific feature that drops the user into the debugger if the key sequence is pressed.  This seems to be present so that the key sequence triggers dropping into the debugger not just from the main window, but also from debugger windows.

This has also created an oddity where pressing the debugger key sequence from a MAME window will cause a "User-initiated break" but pressing it from a debugger window will cause "Internal breakpoint" to be displayed.

However, while this logic has a check to not apply when to a MAME window, it seems to also activate even if MAME itself is not in focus.  This change ensures that a stray debugger sequence in a completely unrelated application won't cause the user to be dropped in the debugger.