formats/all.cpp: Include IPF, 86F disk formats in floptool (#13935)
IPF is included unconditionally by scripts/formats.lua, so the
HAS_FORMATS condition was always false.

86F is also included unconditionally but simply wasn't added to
formats/all.cpp when it was introduced.

I added 86F to the "Container FM/MFM" category for consistency with the
grouping in devices/imagedev/floppy.cpp and moved IPF to the "PC"
category for the same reason.