CD-i: Add Undefined Matte Behavior and fix RGB Transparency  (#13978)
* CD-i: Add Undefined Matte Behavior
Fixes #13976
Greenbook says not to change the MF flag within a line. However, if you do, the behavior is undefined.
Console verification shows that the real console will use these values. The non-spec behavior is explained in a comment.
Additionally, this tidies up the Matte function.

* CD-i: Fix RGB Transparency
This fixes RGB transparency checks.
Previously it was checking the high bit. This would pass when checking the false condition (mode 10) but not on the true condition (mode 2).