CD-i: Fix Cursor Inversion Color and Add External Video Placeholder (#13989)
* CD-i: Fix Cursor Inversion Color
This corrects a slight error in the cursor inversion color. A full brightness color should keep the brightness value, and only invert the RGB. See MCD212 section 7.5

* CD-i: Introduce External Video Input
This replaces the static background color with a function that will switch behavior if the EV flag is set. Currently there is no DVC chip, so this will produce black output.  This fixes #13980 .