CD-i: Add Interlace Graphics (#13997)
* CD-i: Doubles Vertical Resolution

Minimal code to double the vertical resolution for #13231 . The result is a slightly crisper looking image on most monitors due to better upscaling on images are have less oblique pixels.

This was tested minimally and seems stable.

Something closer to a proper interlaced image can be accomplished using the _PA bit, however current screen blanks between frames preventing the parity bit from being able to produce a correct interlace image yet.

* CD-i: Remove excess vertical resolution

* Use 2x clock multiplier

* CD-i: Add Interlaced Graphics

Fixes #13231

* CD-i: Toggle Interlace Effect

When the interlace bit is off, line-duplication is re-enabled. This ensures the interlace effect is only visible when the game declares the output to be interlaced.

* CD-i: Fix interlaced line offset

In some cases, such as during startup, the line length of the interlaced line is impacted. This corrects the offset.

* CD-i: Fix Interlace Clock Speed