apple2: fix regression in lo-res and text rendering
Changes in 8671820 caused the cliprect to be ignored during lo-res
and text rendering.  The resulting over-drawing artifacts were
generally transient and hard to notice, but more obvious in beam-racing
demos (by Deater, French Touch, etc.)

This commit fixes the logic to properly clip again: explicit row
alignment is removed (without re-introducing #5595; an offset cliprect
is still aligned to the correct address and nibble), and the lo-res
memcpy optimization is clipped.