ef9345: scanline generation should not assert/deassert the BUSY flag
The status register's BUSY flag should only be asserted by EF9345
while a user command is pending or being executed. Rendering (i.e.
generation of the next line of pixels) is an independent process
and it should neither set it nor clear it.

The current code, instead, always asserts it at the beginning of
the scanline generation periodic task and deasserts it shortly
afterwards (after a fixed timeout). This causes the real BUSY
value and timer, that were previosly configured as a result of
user commands in ef9345_exec calling set_busy_flag, to be lost.

Figure 10 of the EF9345 datasheet shows that, in real HW,
rendering just delays memory accesses due to user commands by
104 us (=40+24+40). As an approximation, this commit delays the
completion of the current command, irrespective of whether it
involves memory accesses or not.