This change is required in order for BIOS for the AT&T 6300 Plus to work.
The 6300 Plus has external logic to reset the 80286 via the reset pin,
so that it can return to Real Mode from protected mode.  Upon return to
Real Mode, the BIOS code is clearly depending on the fact that SP register
is preserved upon reset (the BIOS does however explicitly load the
SS register after reset, before it is used).

The 8086 and 80286 documentation call out which registers are to be
initialized upon reset, and SS is one of them.  For AX, BX, CX,
DX, BP, SP, SI and DI, the behaviour is not specified.  However, based
on what the 6300 Plus BIOS is doing, it appears the actual hardware
behaviour is to leave them untouched.  Although we can only be fully
confident that the real 286 behaves this way with respect to the SP
register, it's likely that the other registers which do not have
explicitly defined initialization behavior are also not touched during
reset.

Note that this observed behaviour only pertains to reset via the reset
pin, and not to power-on initialization.