flopimg.cpp: add a default implementation for `supports_save` (#14156)
Since the default implementation of `save` is a stub that always fails,
it makes sense to also add a default implementation of `supports_save`
that returns `false`. That way, there's one fewer method that needs to be
implemented for read-only image formats.

Remove the now-redundant implementations in concrete format classes.