formats/ap2_dsk.cpp: refactor a2_16sect_format::save
* extract fragments that would be different for the 13-sector format
  into separate functions;

* extract repeated expressions into variables;

* merge `pos` and `wrap` into a struct;

For simplicity of implementation, this makes a minor behavior change:
when the current sector is of the wrong DOS version, we now start searching
for the next sector at the position after the address postamble, rather than
after the data preamble (IOW, we rewind the current position, which we didn't do
before). This shouldn't make much difference in practice.