formats/ap2_dsk.cpp: remove the RWTS18 format (#14262)
It is completely non-functional. The loading code is commented out,
and the saving code crashes, because it attempts to access 70 tracks
regardless of how many there actually are.

The code has been in this state ever since it's been added in 36ea482a,
so it seems unlikely that it'll improve. Moreover, its utility seems
questionable: I haven't been able to find a single floppy image with an .rti
extension (and there are none listed in the software list). OTOH, for every
game mentioned in the description, there are WOZ images that can be used
instead (except for The Toy Shop, which has DSK images).