g65816: fix IRQ to be level triggered
Followup b3f2535: stop clearing LINE_IRQ when jumping to the IRQ vector.
Properly written IRQ handlers won't notice any difference, but this fixes
the behavior for software which (possibly intentionally) fails to clear
an IRQ source; in that case the IRQ should continue to trigger (probably
leading to a system hang.)

LINE_SO is generally not hooked up, but if it is, it doesn't raise IRQ.

Also delete unused G65816_INT_* macros.