esq5505.cpp: Add VFX-family ROM & EEPROM Cartridge support, and improve floppy support.
This adds vfxcart.h and .cpp, which implement a cartridge slot device
and a matching cartridge image device.

The VFX also uses DUART input 1 for the cartridge presence, just like the VFX-SD.

This PR also fixes recognition of the floppy drive, so a floppy
can now be inserted and read from.

This includes loading the sequencer OS, sound and sequence files,
and successfully playing back sequences and songs.

It also tracks the presence of the disk(image) in the floppy drive to synthesize
the DSKCHG signal that the (VFS)SD(-1) expects, and combines that with the
OTIS IRQ to generace the correct DOCIRQ.

The SD keyboards pulse the ISK_ENABLE & MOTOR_ON signals to detect a disk change.
Where a real floppy drive doesn't have the time to react to such a brief pulse and
the motor remains off, the floppy_image_device immediately starts playing the
floppy motor sound. So we delay MOTOR_ON by 100 microseconds to prevent the simulated
motor from starting from just such a short pulse.