minitel2: Improved ROM list
Different revisions of the Minitel 2 came with slightly different
ROMs. In addition to the ones already listed (Bv4 and Bv9), this
commits adds references to Bv6 and Bv7 too.

The descriptions of the ROMs were updated to mention "France
Telecom" (the "ft_" in their names), who commissioned this Minitel
model.

A generic ROM slot was added too, with the maximum possible size
(64 KiB). This makes it easy use precompiled MAME binaries to
test/debug an homebrew ROM too, with -rom FILENAME.bin