Mark the sd132 as the parent, and the sd1 (21-voice) as the clone.
Per [the documentation](https://docs.mamedev.org/usingmame/commonissues.html#why-is-it-that-some-games-have-the-us-version-as-the-main-set-some-have-japanese-and-some-are-the-world),

> Parent and clone sets are a convenience feature to help keep
> different versions of the same system or software together. The
> decision on which set to make the parent will always be somewhat
> arbitrary, but we do have some guidelines:
>
> - Prefer latest release version
> [ ... ]
> - Prefer most complete version
> [ ... ]

The SD-1/32 is the most complete version of the VFX family, and
definitely of the SD-1 variant of the family.

Arguably, the VFX-SD and VFX could also be marked as clones, but
there's a bigger difference in each one of those steps, so I've
not made that particular arbitrary choice.