apple2e: fix IIc VBL behavior
C019 on the IIc behaves quite differently than the IIe or IIgs;
it does not signify that the beam is in VBL, rather it signifies
that a VBL IRQ has fired.  This happens when the beam reaches VBL,
but only if C05B ENVBL is set, and C019 remains set until manually
cleared via accessing C07X.  This is similar to how C041 and C047
work on the IIgs.

Also fix a logic error in laser_mouse_r() breaking VBL polling.