apple2e: cleanup IIc (#14388)
* apple2e: IIc does not have internal gameio

Followup #14307: the IIc (and clones based on this form factor)
do not have an internal 16-pin gameio port, and the annunciators
are not routed to the external joystick connector, thus devices
such as gizmo and compeyes can't work.

* apple2e: remove c000_iic_w()

Only C078-7F have IIc-specific behavior.  Handle them in common
c000_w, to reduce copy-paste.  Three functional changes:
* Zip Chip C058-5F now work in a non-plus IIc
* Zip Chip C058-5F now side-effect annunciators on the IIe
* C07X now side-effect paddle timers and VBL interrupt on the IIc

* apple2e: collapse c000_iic_r()

Following the pattern of c000_laser_r(), reduce c000_iic_r() to
the minimum set of IIc-specific softswitches and rely on c000_r()
for the rest.  Three functional changes:
* unconnected joysticks now read properly (8021d6b missed the IIc)
* Zip Chip C05B now works the same on the IIc and the IIe
* RDIOUDIS/RDDHIRES now side-effect paddle timers and VBL interrupt

* apple2e: fix IIc VBL behavior

C019 on the IIc behaves quite differently than the IIe or IIgs;
it does not signify that the beam is in VBL, rather it signifies
that a VBL IRQ has fired.  This happens when the beam reaches VBL,
but only if C05B ENVBL is set, and C019 remains set until manually
cleared via accessing C07X.  This is similar to how C041 and C047
work on the IIgs.

Also fix a logic error in laser_mouse_r() breaking VBL polling.

* apple2e: fix IIc C04x readback

Fix C04x to return the floating bus in the low 7 bits, matching
behavior verified on hardware.

* apple2e: remove redundant m_cassette_out

C020 now works identically reading or writing, the same as apple2.