Move the ioports from the esqpanel into the esq5505 device - but still make them call the panel when they change.
This seems to give us something like the best of both worlds: the Layouts can be loaded by the machine driver (esq5505_device), and use relative paths in their inputtags to find the io ports. But the code to handle the changes to the input is in the esqpanel class.