Some refactoring to clarify responsibilities.
The esqpanel_vfx device adds a new static method that will add the IO ports it needs, to the specified owning device, and with change callbacks set to its own methods at the tag specified by the caller.

esq5505_state calls the panel class's method when creating its input ports, passing itself asa the owner,and the tag under which it also declares the panel.

This way, the panel is the only piece of C++ code that has to specify the io port names; and the esq5505_state device is the only place that has to know the tag where the panel lives.

The layouts of course also need to know the names of the io ports, but trhey do not need to know their absolute paths.

I've also added a 'panel_' prefix to the io port names, both for clarity (that these are panel-related io ports) and to reduce the likelihood of name collisions (should the esq5505_state class need to add other IO ports).