apple2gs: fix Zip Chip inconsistencies
Followup #9503: fix the ZipGS enable/disable to work correctly,
and not re-lock.  Also properly initialize the paddle delay,
which has been verified against hardware to work the same way
as the Zip Chip documentation (not reversed.)  Also consistently
side-effect the annunciators.

This fixes errors visible in the ZipGS CDA/CDEV.