apple: slightly improve Zip Chip emulation (#14397)
* apple2gs: fix Zip Chip inconsistencies

Followup #9503: fix the ZipGS enable/disable to work correctly,
and not re-lock.  Also properly initialize the paddle delay,
which has been verified against hardware to work the same way
as the Zip Chip documentation (not reversed.)  Also consistently
side-effect the annunciators.

This fixes errors visible in the ZipGS CDA/CDEV.

* apple: improve Zip Chip C05B

Implement the disable and delay status bits.  Now software can
detect the current enable state of the Zip, so the user's boot-time
preference can be restored, instead of blindly enabling or disabling.