Delete unused code
These functions are unused and appear to be a previous implementation. There are some slight differences in behavior compared to the current used code, such as mono audio not storing duplicate samples for each output device. These differences may be bugs, so storing the unused code does not seem to have clear value.

This deletes the unused code.