Improve the Panel, and add three more views, including a working keyboard.
The panel view is improved after re-measuring on a real keyboard. Colors
have been changed a bit in order to improve contrast.

Three new views have been added:

1. a Full view of the keyboard, including the controls to the left of the
   keyboard, specifically the Pitch-Bend and Modulation Wheels and the
   Patch Select buttons, as well as a representation of the floppy drive
   and the little "storage compartment" above it that is a bit of a
   characteristic shape.

2. a "Compact" view that includes all the controls and a 37-key (3-octave)
   keyboard that may be both legible and playable at the same time, on a
   reasonably-sized touch screen.

3. a "Tablet" view of the front-panel controls, with the display and buttons
   stacked vertically instead of stretched horizontally, as well as the
   "Patch Select" buttons. This view, together with a MIDI keyboard
   controller, this should give users a reasonably workable facsimile of
   the real thing: Keys, Pitch Bend and Modulation wheels on their
   MIDI keyboard controller; all the keyboard-specific controls and the
   Ensoniq-specific Patch Select buttons on the "Tablet" view.

The Full and Compact views also include a keyboard: the full 61 keys and a
reduced 37 keys respectively. These work, tracking pointer events, and
generate not just key-down events with velocity, but also per-key
"pressure" or "aftertouch". Velocity is simulated by position of the
click/touch; Pressure by movement of the pointer on the key while down.
Both are also animated: Velocity is shown as a colour between blue (1)
and green (127), Pressure between yellow (1) and red (127).