Mark the Ensoniq SD-1/32 as working. (#14500)
* es5510.h, esqpump.h: Remove feature::SOUND from imperfect_features().

Empirically, the emulated VFX-family of keyboards sound exactly like
the real ones - I've compared many sounds, and more than that, many
demo songs and sequences played by the SD-1/32 sequencer.

So to me, the es5510 processes, and the esq_5505_5510_pump feeds, the
sound just as they should.

Accordingly I think both should _not_ be marked as having
imperfect sound.

* Update machine flags for esq5505 devices.

Specifically, the sd132 is working! It has working floppy and cartridge,
runs the sequencer OS, plays the sounds it should, plays back sequences
and songs and allows editing them, MIDI works - it's all there.

For some strange reason, the 21-voice keyboards in the same family have
timing issues, so marking those accordingly.

* Mark the sd132 as the parent, and the sd1 (21-voice) as the clone.

Per [the documentation](https://docs.mamedev.org/usingmame/commonissues.html#why-is-it-that-some-games-have-the-us-version-as-the-main-set-some-have-japanese-and-some-are-the-world),

> Parent and clone sets are a convenience feature to help keep
> different versions of the same system or software together. The
> decision on which set to make the parent will always be somewhat
> arbitrary, but we do have some guidelines:
>
> - Prefer latest release version
> [ ... ]
> - Prefer most complete version
> [ ... ]

The SD-1/32 is the most complete version of the VFX family, and
definitely of the SD-1 variant of the family.

Arguably, the VFX-SD and VFX could also be marked as clones, but
there's a bigger difference in each one of those steps, so I've
not made that particular arbitrary choice.