sd1 (21 voice) BIOS 4.xx now correctly detects speed.
This allows the 4.xx BIOS to correctly configure the emulated sd1 for
21 voices and thus play at the correct pitch.

Thus the SD-1 Sequencer OS 4.xx floppies now also apply
to the SD-1.

So this also consolidates the `sd1_flop` and `sd132_flop` software lists
into a single `sd1_flop` list - but still keeps the `vfxsd_flop` list
separate because that is still sufficiently different.

The consolidated `sd1_flop` software list uses
`<sharedfeat name="compatibility" ... />` to allow the different sd1
versions to show only those floppies that apply, respectively:
The sd132 does not show the Sequencer OS 3.00 floppy; the sd1
does not show the demo disks.

Arguably this is not quite right because all the demos would at least
technically work even on the 21-voice SD-1, in particular those for other
keyboards (SQ and KS series) which only use the SD-1's sequencer;
but any SD-1/32 demos would run out of voices and thus sound worse than
they should if played on an SD-1 (21-voice), so it seems a safer bet to
make those only available on the SD-1/32.