i86.cpp: Effective Address calculatation time varies greatly from one (#14632)
CPU variant to the next.  On the 286, it's effectively free, unless
two additions are required, in which case it takes 1 cycle.  On the
8088/8086 on the other hand, it can take up to 12 cycles.  There
needs to be a different set of timings for each CPU variant.
I tested on a 286 system (att6300p) with https://thandor.net/benchmark/36,
and the score went from 29 to 46 - almost a 60% speedup!
Also added a few minor cycle count fixes.