DRAFT: Removing `emu_options` reference from `machine_config`
This change is intended to be a (as the circumstances allow) minimal change required to decouple `emu_options` from `machine_config`, and doesn't address the broader Sisyphean issues of how software list handling is coupled with device loading.  For this reason, there is some grossness done in the name of "doing no harm", but this does succeed in decoupling `machine_config` and `emu_options`

* Because `software_list_device` no longer has access to an `emu_options`, it can no longer parse software lists on demand, and the various call sites that can exercise this code path now have to ensure that software lists are parsed.
* `software_get_default_slot()` now needs the image name from options to be passed.  This was a somewhat intrusive name because some additional grossness needed to be added to `get_default_card_software_hook`

A proximate benefit is now that when `-listxml` is invoked, we don't have to feed `machine_config` an `emu_options`, and the resulting configuration is created without any slots specified.  This removes some of the extraneous `-listxml` items associated with the default configuration.

This change has a known issue that should block merging.  A number of communications devices (e.g. - `m3comm_device`) were grabbing `emu_options` in their constructors to build hostnames.  While it would be straight forward to change these to be lazy loaded, we probably should identify a better way for this pattern to be implemented.