ibm5170: add Doom 2 Germany release
This is not sourced from original media.  I had copies of the files on
my hard disk, and while it matches known-good checksums (eg:
https://doomwiki.org/wiki/DOOM2.WAD), I had to reconstruct the floppy
set myself.  The set was created by running msdos622 in the ct486
driver with a CD-ROM attached.  I put the loose files on an ISO image,
mounted it in the DOS machine, and then formatted floppies and copied
the files to them just as the normal Doom installation sets work.

Due to this nature, I am marking this as a “bad dump” in the hopes
someone may be able to find and dump the original disks.

New working software list item (ibm5170.xml)
--------------------------------------------
DOOM II: Hell on Earth (1.666) (Germany) [chungy]