Now Layout-Lua-script only.
Also changes key state changes from one animation state callback for each key (called for each frame), to explicitly changing the state of only the affected key and only when there is actually somethign to change.

Hardcodes the paths to the image devices in the interest of reducing the number of lines of code. To make this more flexible, could iterate through image devices until it finds one with a matching local tag (either `cart` or `35dd`), but that seems excessive at this point.