minitel2: Implement sound output for DTMF tones and beeps
In the minitel2, sound is generated by a speaker connected to
the TS7514 modem chip. A subset of its features are implemented
in this commit:
 - Mode selection, using MCBC and DTMF lines.
 - Shift register for input control data, using the RTS (clock)
   and PRD (data) lines.
 - Modem register RDTMF, to select what pair of frequencies to
   generate while in DTMF mode.
 - Modem register RWLO, selecting what to play on the speaker. In
   particular, monitoring of the transmitted DTMF tones and beep
   generation have been implemented. Monitoring of TX/RX data
   tones and volume control has NOT been implemented.
 - Modem register RPTF, selecting a filter on the transmitted
   frequencies. Only the filter on DTMF tones (to select only one
   of the two frequencies in the current pair) has been
   implemented.

This subset of features is enough to play all the sounds generated
by the "demov1" ROM and the original ROM (beep on invalid
keypress, DTMF dialing tones).

This commit also simplifies the pre-existing logic to generate a
carrier-detect signal when the line control relay is closed.