CD-i: Fix Softlock in Fun Tower
Fixes #14462

Spinup time of 1 is too fast for real hardware, and in practice causes some softlocks. Spinup time of 6 is arbitrary and chosen as the minimal number which fixes this softlock.

The best possible fix would be to calculate a number dynamically based on the current disc state. A TODO is added as a reminder that this is an arbitrary constant, and not the perfect implementation.

I tested Zelda, Wand of Gamelon, and a few other games, and didn't notice any issues.